Base code
I'll be real some of this code was made on my pc and the rest was made on my laptop. i have not been able to get anything to compile on my laptop yet but was on my pc. so some of the code might have errors as ive yet to compile it in a while and mainly have just been going off online guides on gtk